Install go-cd server from debian repo
This is to fix the behaviour observed when directly downloading a debian
file and installing it via dpkg_package resource. In this situation,
dpkg does not recognize if a newer version than the one installed has
been downloaded. Also, using a debian repo means we're relying on proven
technology to manage packages and versions.